Déplacement des urls/views/admin dans chaque app
