+++ lettings/urls.py
@@ -1,0 +1,8 @@
+from django.urls import path
+
+from . import views
+
+urlpatterns = [
+    path('lettings/', views.lettings_index, name='lettings_index'),
+    path('lettings/<int:letting_id>/', views.letting, name='letting'),
+]
