+++ oc_lettings_site/settings.py
@@ -19,15 +19,16 @@ ALLOWED_HOSTS = []
 
 
 # Application definition
-# TODO: les apps lettings et profiles sont à créer et à ajouter dans INSTALLED_APPS
 INSTALLED_APPS = [
-    'oc_lettings_site.apps.OCLettingsSiteConfig',
     'django.contrib.admin',
     'django.contrib.auth',
     'django.contrib.contenttypes',
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
+    'oc_lettings_site.apps.OCLettingsSiteConfig',
+    'lettings.apps.LettingsConfig',
+    'profiles.apps.ProfilesConfig',
 ]
 
 MIDDLEWARE = [
