--- oc_lettings_site/tests.py
@@ -1,3 +1,0 @@
-# TODO: créer un dossier "tests" en racine du projet avec unit/ , functional/ et integration/ ?!?
-def test_dummy():
-    assert 1
