+++ profiles/admin.py
@@ -1,3 +1,5 @@
 from django.contrib import admin
 
-# Register your models here.
+from .models import Profile
+
+admin.site.register(Profile)
